fixes https://github.com/fabric8io/fabric8/issues/4524
