+++ quickstarts/spring-boot/camel/src/test/java/io/fabric8/tests/integration/KubernetesIntegrationKT.java
@@ -13,6 +13,7 @@
  * implied.  See the License for the specific language governing
  * permissions and limitations under the License.
  */
+
 package io.fabric8.tests.integration;
 
 import io.fabric8.arquillian.kubernetes.Session;
@@ -27,7 +28,7 @@ import org.junit.runner.RunWith;
 import static io.fabric8.kubernetes.assertions.Assertions.assertThat;
 
 @RunWith(Arquillian.class)
-public class AppKubernetesTest {
+public class KubernetesIntegrationKT {
 
     @ArquillianResource
     KubernetesClient client;
